+++ uui/components/datePickers/RangeDatePickerBody.tsx
@@ -1,12 +1,13 @@
 import React, { forwardRef, useState } from 'react';
-import { Dayjs, uuiDayjs } from '../../helpers/dayJsHelper';
 import { cx, IControlled } from '@epam/uui-core';
 import {
     uuiDaySelection, Day, DayProps, RangeDatePickerPresets,
 } from '@epam/uui-components';
 import { FlexCell, FlexRow } from '../layout';
 import { CalendarPresets } from './CalendarPresets';
-import css from './RangeDatePickerBody.module.scss';
+import { StatelessDatePickerBody, StatelessDatePickerBodyValue } from './DatePickerBody';
+
+import { Dayjs, uuiDayjs } from '../../helpers/dayJsHelper';
 import {
     defaultRangeValue, getMonthOnOpen, getWithFrom, getWithTo, uuiDatePickerBodyBase, valueFormat,
 } from './helpers';
@@ -14,7 +15,8 @@ import {
     CommonDatePickerBodyProps,
     RangeDatePickerInputType, RangeDatePickerValue, RangeDatePickerBodyValue, ViewType,
 } from './types';
-import { StatelessDatePickerBody, StatelessDatePickerBodyValue } from './DatePickerBody';
+
+import css from './RangeDatePickerBody.module.scss';
 
 export const uuiRangeDatePickerBody = {
     inRange: 'uui-range-datepicker-in-range',