+++ uui/components/tables/ColumnHeaderDropdown/SortingPanel.tsx
@@ -1,11 +1,11 @@
 import React, { useCallback } from 'react';
-import css from './SortingPanel.module.scss';
-import { ReactComponent as SortIcon } from '@epam/assets/icons/table-sort_asc-outline.svg';
-import { ReactComponent as SortIconDesc } from '@epam/assets/icons/table-sort_desc-outline.svg';
 import { cx, SortDirection } from '@epam/uui-core';
 import { FlexCell } from '../../layout';
-import { i18n } from '../../../i18n';
 import { DropdownMenuButton } from '../../overlays';
+import { i18n } from '../../../i18n';
+import { settings } from '../../../settings';
+
+import css from './SortingPanel.module.scss';
 
 export interface SortingPanelProps {
     sortDirection: SortDirection;
@@ -18,8 +18,18 @@ const SortingPanelImpl: React.FC<SortingPanelProps> = ({ sortDirection, onSort }
 
     return (
         <FlexCell cx={ cx(css.sortingPanelContainer, 'uui-dropdownMenu-body') }>
-            <DropdownMenuButton isActive={ sortDirection === 'asc' } caption={ i18n.pickerFilterHeader.sortAscending } icon={ SortIcon } onClick={ sortAsc } />
-            <DropdownMenuButton isActive={ sortDirection === 'desc' } caption={ i18n.pickerFilterHeader.sortDescending } icon={ SortIconDesc } onClick={ sortDesc } />
+            <DropdownMenuButton
+                isActive={ sortDirection === 'asc' }
+                caption={ i18n.pickerFilterHeader.sortAscending }
+                icon={ settings.dataTable.icons.header.ascSortIcon }
+                onClick={ sortAsc }
+            />
+            <DropdownMenuButton
+                isActive={ sortDirection === 'desc' }
+                caption={ i18n.pickerFilterHeader.sortDescending }
+                icon={ settings.dataTable.icons.header.descSortIcon }
+                onClick={ sortDesc }
+            />
         </FlexCell>
     );
 };